fix: move Container component to own directory
