+++ src/index.js
@@ -18,7 +18,8 @@ export { default as Carousel, CarouselItem } from './Carousel';
 export { default as CheckBox } from './CheckBox';
 export { default as CheckBoxGroup } from './CheckBoxGroup';
 export { default as CloseButton } from './CloseButton';
-export { Col, Row, Container } from './Layout';
+export { default as Container } from './Container';
+export { Col, Row } from './Layout';
 export { default as Collapse } from './Collapse';
 export { default as Collapsible } from './Collapsible';
 export { default as Dropdown, DropdownButton, SplitButton } from './Dropdown';
